requirements.txt
openai
chromadb
tqdm
python-dotenv
requests
beautifulsoup4
flask
flask_cors
